🩹 patch: DRY clients and standardize error handler
