+++ packages/clients/lib/baseClient.ts
@@ -8,7 +8,7 @@ import axios, {
   Method,
 } from 'axios';
 
-import { IOracleOptions } from './options';
+import { IOptions } from './options';
 
 const API_PREFIX = 'api/v1';
 
@@ -36,71 +36,11 @@ export interface AnnouncementResponse {
 /**
  * Oracle client
  */
-export class OracleClient {
-  public opts: IOracleOptions;
+export class BaseClient {
+  public opts: IOptions;
 
-  constructor(opts: IOracleOptions) {
+  constructor(opts: IOptions) {
     this.opts = opts;
-  }
-
-  /**
-   * getAnnouncements
-   *
-   * Get all announcements
-   *
-   * @returns
-   */
-  public async getAnnouncements(
-    offerId?: string,
-    startDate?: number,
-    endDate?: number,
-    page?: number,
-    limit?: number,
-  ): Promise<AnnouncementResponse> {
-    return this.get('/announcements', {
-      offerId,
-      startDate,
-      endDate,
-      page,
-      limit,
-    });
-  }
-
-  /**
-   * getAnnouncementByEventId
-   *
-   * Get Announcement By Event ID
-   *
-   * @param {string} eventId the eventId for the DLC OracleAnnouncement
-   *                         https://github.com/discreetlogcontracts/dlcspecs/blob/master/Messaging.md#oracle_event
-   * @returns {Announcement}
-   */
-  public async getAnnouncementByEventId(
-    eventId: string,
-  ): Promise<Announcement> {
-    return this.get(`/announcements/${eventId}`);
-  }
-
-  /**
-   * postAnnouncementQuantity
-   *
-   * Specify quantity of funds entered for a particular announcement
-   * Note: this function does not modify the DLC OracleAnnouncement
-   *       itself but rather the Oracle internal DB
-   *
-   * @param {string} eventId the eventId for the DLC OracleAnnouncement
-   *                         https://github.com/discreetlogcontracts/dlcspecs/blob/master/Messaging.md#oracle_event
-   * @param {number} quantity the quantity entered in SATS
-   * @returns {Announcement} return modified announcement
-   */
-  public async postAnnouncementQuantity(
-    eventId: string,
-    quantity: number,
-  ): Promise<Announcement> {
-    if (!this.opts.apiKey)
-      throw Error('API Key required for postAnnouncementQuantity');
-
-    return this.post(`/announcements/${eventId}/quantity`, { quantity });
   }
 
   public async request(
@@ -168,19 +108,19 @@ export class OracleClient {
 
   public handleError = (error: AxiosError): void => {
     if (error.code === 'ECONNREFUSED') {
-      this.opts.logger.error(
-        `Error: Could not connect to Oracle Service ${this.opts.uri}`,
-      );
       this.opts.logger.debug(
-        'Make sure the Oracle Service is running and that you are connecting to the correct port',
+        'Make sure the Service is running and that you are connecting to the correct port',
       );
-    } else if (error.response) {
-      this.opts.logger.error(error.response.data);
-      throw new Error(error.response.data as string);
+      throw Error(`Could not connect to Service ${this.opts.uri}`);
     } else if (error.code === 'EPIPE') {
-      this.opts.logger.error(`EPIPE error ${this.opts.uri}`);
+      throw Error(`EPIPE error ${this.opts.uri}`);
+    } else if (error.response) {
+      if (typeof error.response.data === 'string') {
+        throw new Error(error.response.data);
+      } else {
+        throw new Error((error.response.data as any).error as string);
+      }
     } else {
-      this.opts.logger.error(error.message);
       throw new Error(error.message);
     }
   };
