+++ packages/clients/lib/options.ts
@@ -1,9 +1,10 @@
 import { Logger } from '@node-dlc/logger';
 export type ApiPrefix = 'api';
 
-export interface IStrategyOptions {
+export interface IOptions {
   uri: string;
   logger: Logger;
+  apiKey?: string;
   headers?: [string, string][];
   prefix?: string;
 }
